Extracted project metadata to a new Utility class
