+++ test/groovy/org/ods/orchestration/util/ContextSpec.groovy
@@ -2,7 +2,6 @@ package org.ods.orchestration.util
 
 import java.nio.file.Files
 
-import org.apache.http.client.utils.URIBuilder
 import org.ods.services.GitService
 import org.ods.orchestration.service.*
 import org.ods.orchestration.usecase.*
@@ -24,10 +23,10 @@ class ContextSpec extends SpecHelper {
     def createContext(Map<String, Closure> mixins = [:]) {
         Context context = Spy(constructorArgs: [steps])
 
-        if (mixins.containsKey("getGitURLFromPath")) {
-            context.getGitURLFromPath(*_) >> { mixins["getGitURLFromPath"]() }
+        if (mixins.containsKey("loadMetadata")) {
+            context.loadMetadata(*_) >> { mixins["loadMetadata"]() }
         } else {
-            context.getGitURLFromPath(*_) >> { return new URIBuilder("https://github.com/my-org/my-pipeline-repo.git").build() }
+            context.loadMetadata(*_) >> { return new FakeProject(context.steps).loadMetadata(it[0]) }
         }
 
         if (mixins.containsKey("loadJiraData")) {
@@ -474,94 +473,6 @@ class ContextSpec extends SpecHelper {
             [key: "NET-7", status: "DONE"],
             [key: "NET-20", status: "DONE"]
         ]
-    }
-
-    def "get Git URL from path"() {
-        given:
-        def project = new Context(steps)
-
-        def path = "${steps.env.WORKSPACE}/a/b/c"
-        def origin = "upstream"
-
-        when:
-        def result = project.getGitURLFromPath(path, origin)
-
-        then:
-        1 * steps.dir(path, _)
-
-        then:
-        1 * steps.sh({ it.script == "git config --get remote.${origin}.url" && it.returnStdout }) >> new URI("https://github.com/my-org/my-repo.git").toString()
-
-        then:
-        result == new URI("https://github.com/my-org/my-repo.git")
-    }
-
-    def "get Git URL from path without origin"() {
-        given:
-        def project = new Context(steps)
-
-        def path = "${steps.env.WORKSPACE}/a/b/c"
-
-        when:
-        def result = project.getGitURLFromPath(path)
-
-        then:
-        1 * steps.dir(path, _)
-
-        then:
-        1 * steps.sh({ it.script == "git config --get remote.origin.url" && it.returnStdout }) >> new URI("https://github.com/my-org/my-repo.git").toString()
-
-        then:
-        result == new URI("https://github.com/my-org/my-repo.git")
-    }
-
-    def "get Git URL from path with invalid path"() {
-        given:
-        def project = new Context(steps)
-
-        when:
-        project.getGitURLFromPath(null)
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to get Git URL. 'path' is undefined."
-
-        when:
-        project.getGitURLFromPath("")
-
-        then:
-        e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to get Git URL. 'path' is undefined."
-
-        when:
-        steps.env.WORKSPACE = "myWorkspace"
-        def path = "myPath"
-        project.getGitURLFromPath(path)
-
-        then:
-        e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to get Git URL. 'path' must be inside the Jenkins workspace: ${path}"
-    }
-
-    def "get Git URL from path with invalid remote"() {
-        given:
-        def project = new Context(steps)
-
-        def path = "${steps.env.WORKSPACE}/a/b/c"
-
-        when:
-        project.getGitURLFromPath(path, null)
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to get Git URL. 'remote' is undefined."
-
-        when:
-        project.getGitURLFromPath(path, "")
-
-        then:
-        e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to get Git URL. 'remote' is undefined."
     }
 
     def "is triggered by change management process"() {
@@ -1150,7 +1061,7 @@ class ContextSpec extends SpecHelper {
 
     def "load metadata"() {
         when:
-        def result = context.loadMetadata()
+        def result = context.loadMetadata(Context.METADATA_FILE_NAME)
 
         // Verify annotations to the metadata.yml file are made
         def expected = new Yaml().load(new File(Context.METADATA_FILE_NAME).text)
@@ -1166,285 +1077,5 @@ class ContextSpec extends SpecHelper {
 
         then:
         result == expected
-    }
-
-    def "load metadata with invalid file"() {
-        when:
-        context.loadMetadata(null)
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to parse project meta data. 'filename' is undefined."
-    }
-
-    def "load project metadata with non-existent file"() {
-        when:
-        def filename = "non-existent"
-        context.loadMetadata(filename)
-
-        then:
-        def e = thrown(RuntimeException)
-        e.message == "Error: unable to load project meta data. File '${steps.env.WORKSPACE}/${filename}' does not exist."
-    }
-
-    def "load project metadata with invalid id"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-
-        when:
-        metadataFile.text = """
-            name: myName
-        """
-
-        context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to parse project meta data. Required attribute 'id' is undefined."
-
-        cleanup:
-        metadataFile.delete()
-    }
-
-    def "load project metadata with invalid name"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-
-        when:
-        metadataFile.text = """
-            id: myId
-        """
-
-        context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to parse project meta data. Required attribute 'name' is undefined."
-
-        cleanup:
-        metadataFile.delete()
-    }
-
-    def "load project metadata with invalid description"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-
-        when:
-        metadataFile.text = """
-            id: myId
-            name: myName
-            repositories:
-              - id: A
-        """
-
-        def result = context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        result.description == ""
-
-        cleanup:
-        metadataFile.delete()
-    }
-
-    def "load project metadata with undefined repositories"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-
-        when:
-        metadataFile.text = """
-            id: myId
-            name: myName
-        """
-
-        def result = context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        result.repositories == []
-
-        cleanup:
-        metadataFile.delete()
-    }
-
-    def "load project metadata with invalid repository id"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-
-        when:
-        metadataFile.text = """
-            id: myId
-            name: myName
-            repositories:
-              - name: A
-        """
-
-        context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to parse project meta data. Required attribute 'repositories[0].id' is undefined."
-
-        when:
-        metadataFile.text = """
-            id: myId
-            name: myName
-            repositories:
-              - id: A
-              - name: B
-        """
-
-        context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to parse project meta data. Required attribute 'repositories[1].id' is undefined."
-
-        cleanup:
-        metadataFile.delete()
-    }
-
-    def "load project metadata with invalid repository url"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-
-        when:
-        metadataFile.text = """
-            id: myId
-            name: myName
-            repositories:
-              - name: A
-        """
-
-        context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to parse project meta data. Required attribute 'repositories[0].id' is undefined."
-
-        when:
-        metadataFile.text = """
-            id: myId
-            name: myName
-            repositories:
-              - id: A
-              - name: B
-        """
-
-        context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to parse project meta data. Required attribute 'repositories[1].id' is undefined."
-
-        cleanup:
-        metadataFile.delete()
-    }
-
-    def "load project metadata with LeVADocs"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-        Context.METADATA_FILE_NAME = metadataFile.getAbsolutePath()
-
-        metadataFile.text = """
-            id: myId
-            name: myName
-            repositories:
-              - id: A
-                name: A
-                url: http://git.com
-            capabilities:
-              - LeVADocs:
-                  GAMPCategory: 5
-                  templatesVersion: "2.0"
-        """
-
-        when:
-        def result = context.init()
-
-        then:
-        result.getCapability("LeVADocs").GAMPCategory == 5
-        result.getCapability("LeVADocs").templatesVersion == "2.0"
-
-        cleanup:
-        metadataFile.delete()
-    }
-
-    def "load project metadata with LeVADocs capabilities but without templatesVersion"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-        Context.METADATA_FILE_NAME = metadataFile.getAbsolutePath()
-
-        metadataFile.text = """
-            id: myId
-            name: myName
-            repositories:
-              - id: A
-                name: A
-                url: http://git.com
-            capabilities:
-              - LeVADocs:
-                  GAMPCategory: 5
-        """
-
-        when:
-        def result = context.init()
-
-        then:
-        result.getCapability("LeVADocs").templatesVersion == "1.0"
-
-        cleanup:
-        metadataFile.delete()
-    }
-
-    def "load project metadata with multiple LeVADocs capabilities"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-        metadataFile.text = """
-            id: myId
-            name: myName
-            repositories:
-              - id: A
-                name: A
-                url: http://git.com
-            capabilities:
-              - LeVADocs:
-                  GAMPCategory: 1
-              - LeVADocs:
-                  GAMPCategory: 3
-        """
-
-        when:
-        context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Error: unable to parse project metadata. More than one LeVADoc capability has been defined."
-
-        cleanup:
-        metadataFile.delete()
-    }
-
-    def "load project metadata with LeVADocs capabilities but without GAMPCategory"() {
-        given:
-        def metadataFile = Files.createTempFile("metadata", ".yml").toFile()
-        metadataFile.text = """
-            id: myId
-            name: myName
-            repositories:
-              - id: A
-                name: A
-                url: http://git.com
-            capabilities:
-              - LeVADocs:
-        """
-
-        when:
-        context.loadMetadata(metadataFile.getAbsolutePath())
-
-        then:
-        def e = thrown(IllegalArgumentException)
-        e.message == "Error: LeVADocs capability has been defined but contains no GAMPCategory."
-
-        cleanup:
-        metadataFile.delete()
     }
 }
